Publishing rewrite
