+++ killua/cogs/cards.py
@@ -14,6 +14,8 @@ from PIL import Image, ImageFont, ImageDraw
 import io
 import aiohttp
 from killua.functions import custom_cooldown, blcheck
+from killua.classes import User, Card, CardLimitReached, CardNotFound
+from killua.constants import ALLOWED_AMOUNT_MULTIPLE, FREE_SLOTS, DEF_SPELLS, VIEW_DEF_SPELLS, INDESTRUCTABLE, PRICES, BOOK_PAGES
 
 with open('config.json', 'r') as config_file:
 	config = json.loads(config_file.read())
@@ -25,25 +27,6 @@ items = db['items']
 guilds = db['servers']
 general = cluster['general']
 shop = general['shop']
-
-###################TODO TODO TODO####################
-'''
-             TO-DOs
-
-1) k!book showing the cards you have in book format like in the anime |Done|
-
-2) k!hunt Hunting for monsters |Done|
-
-3) k!shop you can buy cards, daily offers, some stuff |Done|
-
-4) k!sell sell a card |Done|
-
-5) k!give modify to work with items |Done|
-
-6) Add support for all spell cards usable with k!use |Done (24/24)|
-'''
-###################TODO TODO TODO####################
-
 
 #Data structure for card system:
 
@@ -71,507 +54,6 @@ example = {
 
 cached_cards = {}
 
-ALLOWED_AMOUNT_MULTIPLE = 3
-FREE_SLOTS = 40
-DEF_SPELLS= [1003, 1004, 1019]
-VIEW_DEF_SPELLS = [1025]
-INDESTRUCTABLE = [1026]
-
-PRICES:dict = {
-    'S': 10000,
-    'A': 5000,
-    'B': 3000,
-    'C': 1500,
-    'D': 800,
-    'E': 500,
-    'F': 200,
-    'G': 100
-}
-
-BOOK_PAGES = [
-"""
-📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖
-
-
-A beginners guide of the greed island card system
-
-
-📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖📖
-""",
-f"""If you are not familiar with how this works in the anime:
-The main goal of the game is to obtain all 100 cards in the book. How hard it is to
-obtain a card is determined by it's **rank**. You can find it on the top right of the 
-card. Next to it to the right is a number. This number times {ALLOWED_AMOUNT_MULTIPLE} is the 
-maximum amount of those cards to exist globally. If that limit is exceeded you can't 
-obtain any more cards unless someone looses one of their copies which means one 
-other person can obtain the card again""",
-
-f"""On the top left, you can see the card number. Typically, spell cards have a number
-one thousand and ... and item cards have a number less than 100. Cards with a 
-number below 100 count towards your goal of colleting all 100 **restricted slot** 
-cards. When you obtain a card which has an id below 100 but you already have one 
-in your restricted slots, or the card id is above 100, the card comes into your 
-**free slots**. You can have a maximum of {FREE_SLOTS} cards in your free slots""",
-
-"""I have mentioned before that there are **spell cards**. You can use them to steal 
-cards from other users, gamble and a lot more. To use a spell card, use 
-`k!use <card> <arguments>`. Some spell cards only work in a **short range**. 
-In discord terms that means that the target must have send a message
-recently in the channel the command is used in. You can also use permament 
-spell cards to protect yourself from others or tranform cards into fakes.""",
-
-"""A word about **fakes**. The main usecase I see them as is bait. You can't 
-sell fakes, you can't use them and they don't count towards the 100 card goal. 
-If you want to swap out a fake in your album with a real card in your free slots 
-or the other way around, use `k!swap <card_id>`. If you want to get rid of a fake, 
-make sure it's in your free slots and discard it with `k!discard <card_id>`""",
-
-"""You have reached the end of the introduction!
-
-Now it's time for you to explore the world of cards, steal, collect, form alliances and so on. 
-
-Do you want to add a card to the game you have a good idea for? That is possible, if you can make the
-card design with image we will be happy to have a look at your idea!
-
-Have fun hunters
-- The Gamemaster"""
-]
-
-class CardNotFound(Exception):
-    def __init__(self, msg=None):
-        super().__init__()
-        self.text = msg
-
-class NotInPossesion(Exception):
-    def __init__(self, msg=None):
-        super().__init__()
-        self.text = msg
-
-class OnlyFakesFound(Exception):
-    def __init__(self, msg=None):
-        super().__init__()
-        self.text = msg
-
-class CardLimitReached(Exception):
-    def __init__(self, msg=None):
-        super().__init__()
-        self.text = msg
-    
-
-class Card():
-    """This class makes it easier to access card information"""
-    def __init__(self, card_id:int):
-        card = items.find_one({'_id': card_id})
-        if card is None:
-            raise CardNotFound
-        
-        self.id:int = card['_id']
-        self.name:str = card['name']
-        self.image_url:str = card['Image']
-        self.owners:list = card['owners']
-        self.description:str = card['description']
-        self.emoji:str = card['emoji']
-        self.rank:str = card['rank']
-        self.limit:int = card['limit']
-        try:
-            self.type:str = card['type']
-        except KeyError:
-            items.update_one({'_id': self.id}, {'$set':{'type': 'normal'}})
-            self.type = 'normal'
-
-        if card_id > 1000 and not card_id == 1217: # If the card is a spell card it has two additional properties
-            self.range:str = card['range']
-            self.cls:list = card['class']
-
-    def add_owner(self, user_id:int):
-        """Adds an owner to a card entry in my db. Only used in Card().add_card()"""
-        self.owners.append(user_id)
-        items.update_one({'_id': self.id}, {'$set': {'owners': self.owners}})
-        return
-
-    def remove_owner(self, user_id:int):
-        """Removes an owner from a card entry in my db. Only used in Card().remove_card()"""
-        self.owners.remove(user_id)
-        items.update_one({'_id': self.id}, {'$set': {'owners': self.owners}})
-        return
-
-    def __eq__(self, other):
-        """"This function isn't used but I thought it would be nice to implement"""
-        if isinstance(other, Card): # Checking if the other card is a Card object
-            if self.id == other.id:
-                return True
-            else:
-                return False
-        if isinstance(other, int): # Checking if just the card id was passed
-            if self.id == other:
-                return True
-            else: 
-                return False
-        if isinstance(other, list): # Checking if it's passed like it would be in an inventory [int, {"fake": bool}]
-            if self.id == other[0]:
-                return True
-            else:
-                return False
-        if isinstance(other, dict): # Checking if it's passed from a pymongo items.find_one({"_id": int})
-            if self.id == other['_id']:
-                return True
-            else:
-                return False
-
-class User():
-    """This class allows me to handle a lot of user related actions more easily"""
-    def __init__(self, user_id:int):
-        user = teams.find_one({'id': user_id})
-
-        self.id:int = user_id
-
-        if user is None:
-            print('Yes')
-            self.add_empty(self.id, False)
-            user = teams.find_one({'id': user_id})
-
-        if not 'cards' in user or not 'met_user' in user:
-            self.add_empty(self.id)
-            user = teams.find_one({'id': user_id})
-
-        
-        self.jenny:int = user['points']
-        self.daily_cooldown = user['cooldowndaily']
-        self.met_user:list = user['met_user']
-        self.effects:dict = user['cards']['effects']
-        self.rs_cards:list = user['cards']['rs']
-        self.fs_cards:list = user['cards']['fs']
-        self.all_cards:list = [*self.fs_cards, *self.rs_cards]
-
-    @staticmethod
-    def remove_all():
-        """Removes all cards etc from every user. Only used for testing"""
-        start = datetime.now()
-        user = list()
-        cards = list()
-        for u in [x for x in teams.find()]:
-            if 'cards' in u:
-                user.append(u['id'])
-        teams.update_many({'$or': [{'id': x} for x in user]}, {'$set': {'cards': {'rs': [], 'fs': [], 'effects': {}}, 'met_user': []}})
-
-        for c in [x for x in items.find()]:
-            if len(c['owners']) > 0:
-                cards.append(c['_id'])
-        items.update_many({'$or': [{'id': x} for x in cards]}, {'$set': {'owners': []}})
-
-        return f"Removed all cards from {len(user)} user{'s' if len(user) > 1 else ''} and all owners from {len(cards)} card{'s' if len(cards) > 1 else ''} in {(datetime.now() - start).seconds} second{'s' if (datetime.now() - start).seconds > 1 else ''}"
-
-    @classmethod
-    def is_registered(cls, user_id:int):
-        """Checks if the "cards" dictionary is in the database entry of the user"""
-        u = teams.find_one({'id': user_id})
-        if u is None:
-            return False
-        if not 'cards' in u:
-            return False
-
-        return True   
-
-    @classmethod # The reason for this being a classmethod is that User(user_id) automatically calls this function, 
-    # so while I will also never use this, it at least makes more sense
-    def add_empty(self, user_id, cards:bool=True):
-        """Can be called when the user does not have an entry to make the class return empty objects instead of None"""
-        if cards:
-            return teams.update_one({'id': user_id}, {'$set': {'cards': {'rs': [], 'fs': [], 'effects': {}}, 'met_user': []}})  
-        else:
-            return teams.insert_one({'id': user_id, 'points': 0, 'badges': [], 'cooldowndaily': '','cards': {'rs': [], 'fs': [], 'effects': {}}, 'met_user': []}) 
-        
-    def has_rs_card(self, card_id:int, fake_allowed:bool=True):
-        """Checking if the user has a card specified in their restricted slots"""
-        if card_id in [x[0] for x in self.rs_cards]:
-            if fake_allowed is False and True in [x[1] for x in self.rs_cards if x[0] == card_id]: #Not pretty but should work
-                return False
-            return True
-        else:
-            return False
-
-    def has_fs_card(self, card_id:int, fake_allowed:bool=True):
-        """Checking if the user has a card specified in their free slots"""
-        fs = self.fs_cards
-        if card_id in [x[0] for x in fs]:
-            if fake_allowed is False:
-                if True in [x[1]["fake"] for x in fs if x[0] == card_id]:
-                    for x in [x for x in fs if x[0] == card_id and x[1]['fake'] == True]:
-                        fs.remove([x[0], {"fake": True, "clone": x[1]["clone"]}])
-                    if len([x for x in fs if x[0] == card_id]) == 0:
-                        return False
-            return True
-        else:
-            return False
-
-    def has_any_card(self, card_id:int, fake_allowed:bool=True):
-        """Checks if the user has the card"""
-        total = self.all_cards
-        if card_id in [x[0] for x in total]:
-            if fake_allowed is False:
-                if True in [x[1]["fake"] for x in total if x[0] == card_id]:
-                    for x in [x for x in total if x[0] == card_id and x[1]['fake'] == True]:
-                        total.remove([x[0], {"fake": True, "clone": x[1]["clone"]}])
-                    if len([x for x in total if x[0] == card_id]) == 0:
-                        return False
-            return True
-        else:
-            return False
-
-    def remove_jenny(self, amount:int):
-        """Removes x Jenny from a user"""
-        if self.jenny < amount:
-            raise Exception('Trying to remove more Jenny than the user has')
-        teams.update_one({'id': self.id}, {'$set': {'points': self.jenny - amount}})
-        return
-
-    def add_jenny(self, amount:int):
-        """Adds x Jenny to a users balance"""
-        teams.update_one({'id': self.id}, {'$set': {'points': self.jenny + amount}})
-        return
-
-    def set_jenny(self, amount:int):
-        """Sets the users jenny to the specified value. Only used for testing"""
-        teams.update_one({'id': self.id}, {'$set': {'points': amount}})
-        return
-
-    def remove_card(self, card_id:int, remove_fake:bool=None, payed:bool=False, restricted_slot:bool=None, clone:bool=False):
-        """Removes a card from a user"""
-        card = Card(card_id)
-        if self.has_any_card(card_id) is False:
-            raise NotInPossesion('This card is not in possesion of the specified user!')
-
-        def fake_check(card_list:list):
-            indx = [x for x in card_list if x[0] == card_id]
-            (indx.remove([card_id, {"fake": True, "clone": clone}]) for i in indx if card_list[i]['fake'] == True)
-            if len(indx) == 0:
-                return False
-            else:
-                return True
-
-        def rc(fake:bool, restricted_slot:bool):
-
-            if restricted_slot is False or (self.has_fs_card(card_id) and remove_fake is None):
-                #Honestly I am not sure if what I am doing in this whole function works for all usecases it's intended to
-                self.fs_cards.remove([card_id, {'fake': fake, "clone": clone}])
-            elif restricted_slot is True:
-                self.rs_cards.remove([card_id, {'fake': fake, "clone": clone}])
-            if fake is False:
-                card.remove_owner(self.id)
-            teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': self.rs_cards, 'fs': self.fs_cards, 'effects': self.effects}}})
-            return [card_id, {"fake": fake, "clone": clone}]
-
-        def fake():
-            if remove_fake is None:
-                c = []
-                if restricted_slot is True:
-                    for x in self.rs_cards:
-                        if x[0] == card_id:
-                            c.append(x)
-                elif restricted_slot is False:
-                    for x in self.rs_cards:
-                        if x[0] == card_id:
-                            c.append(x)
-                else:
-                    if self.has_fs_card(card_id):
-                        random_fake = random.choice([x[1]['fake'] for x in self.fs_cards if x[0] == card_id])
-                    else:
-                        random_fake = random.choice([x[1]['fake'] for x in self.rs_cards if x[0] == card_id])
-                    return random_fake
-                return random.choice(c)[1]["fake"]
-            elif remove_fake is False:
-                return False
-            elif remove_fake is True:
-                return True
-
-        if self.has_fs_card(card_id) is False:
-            if fake_check(self.rs_cards) is False and remove_fake is False:
-                raise OnlyFakesFound('The user has no card with this id that is not a fake')
-            else:
-                if not restricted_slot: # This is needed if we want to force to take a card from the restricted slots
-                    return rc(fake(), True)
-                else:
-                    return rc(fake(), restricted_slot)
-        else:
-            if fake_check(self.fs_cards) is False and remove_fake is False:
-                if fake_check(self.rs_cards) is False and remove_fake is False:
-                    raise OnlyFakesFound('The user has no card with this id that is not a fake')
-                if not restricted_slot:
-                    return rc(fake(), True)
-                else:
-                    return rc(fake(), restricted_slot)
-            else:
-                if not restricted_slot:
-                    return rc(fake(), False)
-                else:
-                    return rc(fake(), restricted_slot)
-
-        if payed is not False:
-            self.add_jenny(PRICES[card.rank])
-
-        return 
-
-    def add_card(self, card_id:int, fake:bool=False, clone:bool=False):
-        """Adds a card to the the user"""
-        card = Card(card_id)
-
-        def ac(restricted_slot:bool=False):
-            if restricted_slot is False:
-                self.fs_cards.append([card_id, {'fake': fake, 'clone': clone}])
-            elif restricted_slot is True:
-                self.rs_cards.append([card_id, {'fake': fake, 'clone': clone}])
-            if fake is False:
-                card.add_owner(self.id)
-            teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': self.rs_cards, 'fs': self.fs_cards, 'effects': self.effects}}})
-
-        if self.has_rs_card(card.id) is False:
-            if card_id < 100:
-                ac(True)
-                return
-            return ac()
-
-        if len(self.fs_cards) >= FREE_SLOTS:
-            raise CardLimitReached('User reached card limit for free slots')
-
-        ac()
-
-    def count_card(self, card_id:int, including_fakes:bool=True):
-        "Counts how many copies of a card someone has"
-        card = Card(card_id)
-        card_amount = 0
-        if including_fakes is True:
-            rs_cards = [x[0] for x in self.rs_cards]
-            fs_cards = [x[0] for x in self.fs_cards]
-        else:
-            rs_cards = [x[0] for x in self.rs_cards if x[1]['fake'] == False]
-            fs_cards = [x[0] for x in self.fs_cards if x[1]['fake'] == False]
-
-        for x in [*rs_cards, *fs_cards]:
-            if x == card.id:
-                card_amount = card_amount+1
-
-        return card_amount
-
-    def add_multi(self, *args):
-        """The purpose of this function is to be a faster alternative when adding multiple cards than for loop with add_card"""
-        fs_cards = list()
-        rs_cards = list()
-
-        def fs_append(item:list):
-            if len([*self.fs_cards, *fs_cards]) >= 40:
-                    return fs_cards
-            fs_cards.append(item)
-            return fs_cards
-
-        if len(args) == 1: # I might just pass all items in a list
-            args = args[0]
-
-        for item in args:
-            if item[1]["fake"] is False:
-                Card(item[0]).add_owner(self.id)
-            if item[0] < 99:
-                if not self.has_rs_card(item[0]):
-                    if not item[0] in [x[0] for x in rs_cards]:
-                        rs_cards.append(item)
-                    else:
-                        fs_cards = fs_append(item)
-                else:
-                    fs_cards = fs_append(item)
-            else:
-                fs_cards = fs_append(item)
-
-        teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': [*self.rs_cards, *rs_cards], 'fs': [*self.fs_cards, *fs_cards], 'effects': self.effects}}})
-
-    def has_defense(self):
-        """Checks if a user holds on to a defense spell card"""
-        for x in DEFENSE_SPELLS:
-            if x in [x[0] for x in self.fs_cards]:
-                if self.has_any_card(x, False):
-                    return True
-        return False
-
-    def swap(self, card_id:int): 
-        """Swaps a card from the free slots with one from the restricted slots. Usecase: swapping fake and real card"""
-
-        if (True in [x[1]['fake'] for x in self.rs_cards if x[0] == card_id] and False in [x[1]['fake'] for x in self.fs_cards if x[0] == card_id]):
-            r = self.remove_card(card_id, remove_fake=True, restricted_slot=True)
-            r2 = self.remove_card(card_id, remove_fake=False, restricted_slot=False)
-            self.add_card(card_id, False, r[1]["clone"])
-            self.add_card(card_id, True, r2[1]["clone"])
-            return
-
-        if (True in [x[1]['fake'] for x in self.fs_cards if x[0] == card_id] and False in [x[1]['fake'] for x in self.rs_cards if x[0] == card_id]):
-            r = self.remove_card(card_id, remove_fake=True, restricted_slot=False)
-            r2 = self.remove_card(card_id, remove_fake=False, restricted_slot=True)
-            self.add_card(card_id, True, r[1]["clone"])
-            self.add_card(card_id, False, r2[1]["clone"])
-            return
-
-        return False # Returned if the requirements haven't been met
-            
-
-    def add_effect(self, effect, value): 
-        """Adds a card with specified value, easier than checking for appropriate value with effect name"""
-        l = self.effects
-        l[effect] = value
-        teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': self.rs_cards, 'fs': self.fs_cards, 'effects': l}}})
-
-    def remove_effect(self, effect):
-        """Remove effect provided"""
-        l = self.effects
-        l.pop(effect, None)
-        teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': self.rs_cards, 'fs': self.fs_cards, 'effects': l}}})
-
-    
-    def has_effect(self, effect):
-        """Checks if a user has an effect and returns what effect if the user has it"""
-        if effect in self.effects:
-            return True, self.effects[effect]
-        else:
-            return False, None
-
-    def add_met_user(self, user_id:int):
-        """Adds a user to a "previously met" list which is a parameter in some spell cards"""
-        if not user_id in self.met_user:
-            self.met_user.append(user_id)
-            teams.update_one({'id': self.id}, {'$set': {'met_user': self.met_user}})
-
-    def has_met(self, user_id:int):
-        """Checks if the user id provided has been met by the self.id user"""
-        if user_id in self.met_user:
-            return True
-        else:
-            return False
-
-    def nuke_cards(self, t='all'):
-        """A function only intended to be used by bot owners, not in any actual command, that's why it returns True, so the owner can see if it suceeded"""
-        if t == 'all': 
-            for card in [x[0] for x in self.all_cards]:
-                try:
-                    Card(card).remove_owner(self.id)
-                except:
-                    pass
-            teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': [], 'fs': [], 'effects': {}}}})
-            return True
-        if t == 'fs':
-            for card in [x[0] for x in self.fs_cards]:
-                try:
-                    Card(card).remove_owner(self.id)
-                except:
-                    pass
-            teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': self.rs_cards, 'fs': [], 'effects': self.effects}}})
-            return True
-        if t == 'rs':
-            for card in [x[0] for x in self.rs_cards]:
-                try:
-                    Card(card).remove_owner(self.id)
-                except:
-                    pass
-            teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': [], 'fs': self.fs_cards, 'effects': self.effects}}})
-            return True
-        if t == 'effects':
-            teams.update_one({'id': self.id}, {'$set': {'cards': {'rs': self.rs_cards, 'fs': self.fs_cards, 'effects': {}}}})
-            return True
 
 class Cards(commands.Cog):
 
@@ -581,6 +63,7 @@ class Cards(commands.Cog):
         
     @tasks.loop(hours=6)
     async def shop_update(self):
+        return
         #There have to be 4-5 shop items, inserted into the db as a list with the card numbers
         #the challange is to create a balanced system with good items rare enough but not too rare
         shop_items:list = []
@@ -662,7 +145,7 @@ class Cards(commands.Cog):
 
         await ctx.send(embed=embed)
 
-    @custom_cooldown(7400)
+    @custom_cooldown(5)
     @commands.command()
     async def buy(self, ctx, item:int):
         #h Buy a card from the shop with this command
@@ -710,7 +193,7 @@ class Cards(commands.Cog):
         user.remove_jenny(price) #Always putting substracting points before giving the item so if the payment errors no iten is given
         return await ctx.send(f'Sucessfully bought card number `{card.id}` {card.emoji} for {price} Jenny. Check it out in your inventory with `{self.client.command_prefix(self.client, ctx.message)[2]}book`!')
 
-    @custom_cooldown(30)
+    @custom_cooldown(10)
     @commands.command()
     async def sell(self, ctx, item:int, amount=1):
         #h Sell any amount of cards you own
@@ -732,7 +215,7 @@ class Cards(commands.Cog):
         if in_possesion < amount:
             return await ctx.send(f'Seems you don\'t own enough copis of this card. You own {in_possesion} cop{"y" if in_possesion == 1 else "ies"} of this card')
         
-        await ctx.send(f'You will recieve {PRICES[card.rank]*amount} Jenny for selling {"this card" if amount == 1 else "those cards"}, do you want to proceed? **[y/n]**')
+        await ctx.send(f'You will recieve {int((PRICES[card.rank]*amount)/10)} Jenny for selling {"this card" if amount == 1 else "those cards"}, do you want to proceed? **[y/n]**')
 
         def check(msg):
             return msg.author.id == ctx.author.id and (msg.content.lower() in ['y', 'n'])
@@ -751,8 +234,8 @@ class Cards(commands.Cog):
             else:
                 for i in range(amount):
                     user.remove_card(item, False)
-                user.add_jenny(PRICES[card.rank]*amount)
-                await ctx.send(f'Sucessfully sold {amount} cop{"y" if amount == 1 else "ies"} of card number {item} for {PRICES[card.rank]*amount} Jenny!')
+                user.add_jenny(int((PRICES[card.rank]*amount)/10))
+                await ctx.send(f'Sucessfully sold {amount} cop{"y" if amount == 1 else "ies"} of card number {item} for {int((PRICES[card.rank]*amount)/10)} Jenny!')
 
     @custom_cooldown(2)
     @commands.command()
@@ -1416,6 +899,7 @@ async def book_paginator(self, ctx, page, msg:discord.Message=None, first_time=F
         'title': 'Introduction booklet',
         'description': BOOK_PAGES[page-1],
         'color': 0x1400ff,
+        'image': {'url': 'https://cdn.discordapp.com/attachments/759863805567565925/834794115148546058/image0.jpg'},
         'footer': {'text': f'Page {page}/{len(BOOK_PAGES)}'}
     })
