+++ killua/cogs/events.py
@@ -1,160 +1,113 @@
+import json
+import dbl
 import discord
-from datetime import datetime, date, timedelta
-import time
+from datetime import datetime
 from discord.utils import find
 from discord.ext import commands, tasks
-import pymongo
-from pymongo import MongoClient
-import dbl
-import json
-from json import loads
-from killua.functions import custom_cooldown, blcheck, p
+from killua.functions import p
+from killua.classes import User, Guild
+from killua.constants import PREMIUM_ROLES
 with open('config.json', 'r') as config_file:
 	config = json.loads(config_file.read())
 
-premium = {
-    759891477269839926 : "one_star_hunter",
-    798279223344889957 : "two_star_hunter",
-    798279346389254174 : "triple_star_hunter",
-    769622564648648744 : "server_booster"
-}
-
-cluster = MongoClient(config['mongodb'])
-db = cluster['Killua']
-server = db['guilds']
-teams = db['teams']
-generaldb = cluster['general']
-blacklist = generaldb['blacklist']
-pr = generaldb['presence']
-
 class Events(commands.Cog):
-  
-  def __init__(self, client):
-    self.client = client
-    self.token = config['dbl']
-    self.dblpy = dbl.DBLClient(self.client, self.token, autopost=True)
-    self.status.start()
+
+    def __init__(self, client):
+        self.client = client
+        self.token = config['dbl']
+        self.dblpy = dbl.DBLClient(self.client, self.token, autopost=True)
+        self.status.start()
+
+    @commands.Cog.listener()
+    async def on_ready(self):
+        print('------')
+        print('Logged in as: ' + self.client.user.name + f" (ID: {self.client.user.id})")
+        print('------')
+        self.client.startup_datetime = datetime.now()
     
-  @commands.Cog.listener()
-  async def on_ready(self):
-    print('------')
-    print('Logged in as: ' + self.client.user.name + f" (ID: {self.client.user.id})")
-    print('------')
-    self.client.startup_datetime = datetime.now()
+    @tasks.loop(hours=12)
+    async def status(self):
+        await p(self)
 
-  @tasks.loop(hours=12)
-  async def status(self):
-    await p(self)
-
-  @status.before_loop
-  async def before_status(self):
-      await self.client.wait_until_ready()
-
-  @commands.Cog.listener()
-  async def on_guild_join(self, guild):
-    #Changing the status
-    await p(self)
-    results = server.find({'id': guild.id})
+    @status.before_loop
+    async def before_status(self):
+        await self.client.wait_until_ready()
     
-    for result in results:
-      ID = result
-    #Inserting the guild in the databse if it doesn't exist
-    try:
-      prefix = ID['prefix']#This is a dumb way to do it and will be optimised when I have time
-    except UnboundLocalError:
-      server.update_many({'id': guild.id},{'$set':{'points': 0,'items': '','badges': [], 'prefix': 'k!'}}, upsert=True)
-      prefix = 'k!'
+    @commands.Cog.listener()
+    async def on_guild_join(self, guild):
+        #Changing the status
+        await p(self)
+        g = Guild(guild)
       
-    general = find(lambda x: x.name == 'general',  guild.text_channels)
-    if general and general.permissions_for(guild.me).send_messages:
-        embed = discord.Embed.from_dict({
-            'title': 'Hello {}!'.format(guild.name),
-            'description': f'Hi, my name is Killua, thank you for choosing me! \n\nTo get some info about me, use `{prefix}info`\n\nTo change the server prefix, use `{prefix}prefix <new prefix>` (you need administrator perms for that\n\nFor more commands, use `{prefix}help` to see every command\n\nPlease consider leaving feeback with `k!fb` as this helps me improve Killua',
-            'color': 0x1400ff
-        })
-        await general.send(embed=embed)
-  
-  @commands.Cog.listener()
-  async def on_command_error(self, ctx, error):
-    #This handels the k!bug cooldown
-    if isinstance(error, discord.ext.commands.CommandOnCooldown):
-      m, s = divmod(round(ctx.command.get_cooldown_retry_after(ctx)), 60)
+        general = find(lambda x: x.name == 'general',  guild.text_channels)
+        if general and general.permissions_for(guild.me).send_messages:
+            embed = discord.Embed.from_dict({
+                'title': 'Hello {}!'.format(guild.name),
+                'description': f'Hi, my name is Killua, thank you for choosing me! \n\nTo get some info about me, use `{prefix}info`\n\nTo change the server prefix, use `{prefix}prefix <new prefix>` (you need administrator perms for that\n\nFor more commands, use `{prefix}help` to see every command\n\nPlease consider leaving feeback with `k!fb` as this helps me improve Killua',
+                'color': 0x1400ff
+            })
+            await general.send(embed=embed)
 
-      await ctx.send(f'Wait {m:02d} minutes and {s:02d} seconds before using the command again, thank you for helping to improve killua :3')
-    else:
-      guild = ctx.guild.id if ctx.guild else "dm channel with "+ ctx.author.id
-      command = ctx.command.name if ctx.command else "Error didn't occur during a command"
-      print('------------------------------------------')
-      print(f'An error occured\nGuild id: {guild}\nCommand name: {command}\nError: {error}')
-      print('------------------------------------------')
-
-
-        
-  @commands.Cog.listener()
-  async def on_connect(self):
-    #Changing Killua's status
-    await p(self)
+    @commands.Cog.listener()
+    async def on_connect(self):
+        #Changing Killua's status
+        await p(self)
     
 
-  @commands.Cog.listener()
-  async def on_guild_remove(self, guild):
-    #Changing Killua's status
-    await p(self)
-    server.delete_one({'id': guild.id})
+    @commands.Cog.listener()
+    async def on_guild_remove(self, guild):
+        #Changing Killua's status
+        await p(self)
+        Guild(guild.id).delete()
 
-  @commands.Cog.listener()
-  async def on_member_update(self, before, after):
-    try:
-      if not before.guild.id == 715358111472418908:
-        return
-    except:
-      s = list(premium.keys())
-      b = []
-      a = []
-      for role in before.roles:
-        b.append(role.id)
-      for role in after.roles:
-        a.append(role.id)
+    @commands.Cog.listener()
+    async def on_member_update(self, before, after):
+        # The purpose of this function is to kinda automate premium badges
+        try:
+            if not before.guild.id == 715358111472418908:
+                return
+        except:
+            return
+        s = list(PREMIUM_ROLES.keys())
 
-      common_elements = set(s).intersection(b)
-      if common_elements:
-        for element in common_elements:
-          if element in s and not element in a:
-            await remove_premium(before, element)
+        common_elements = set(s).intersection([x.id for x in before.roles])
+        if common_elements:
+            for element in common_elements:
+                if element in s and not element in [x.id for x in after.roles]:
+                    User(after.user.id).remove_badge(PREMIUM_ROLES[element], premium=True)
+                    return await after.user.send('Your Killua premium subscription ran out! Your premium permissions have sadly been taken :c. I hope you enjoyed your time as Killua supporter. (If you see this message after this you left Killuas server please rejoin to get Premium permissions again as you need to be on the server for that)')
 
-      common_elements = set(s).intersection(a)
-      if common_elements:
-        for element in common_elements:
-          if element in s and not element in b:
-            await add_premium(after, element) 
-  
-async def remove_premium(member:discord.Member, s_id:int):
-  user = teams.find_one({'id': member.id})
-  badges = user['badges']
-  badges.remove('premium')
-  badges.remove(premium[s_id])
-  teams.update_one({'id': member.id}, {'$set': {'badges': badges}})
-  try:
-    await member.send('Your Killua premium subscription ran out! Your premium permissions have sadly been taken :c. I hope you enjoyed your time as Killua supporter. (If you see this message after this you left Killuas server please rejoin to get Premium permissions again as you need to be on the server for that)')
-  except:
-    pass
-  return
+        common_elements = set(s).intersection([x.id for x in after.roles])
+        if common_elements:
+            for element in common_elements:
+                if element in s and not element in [x.id for x in before.roles]:
+                    User(after.user.id).add_badge(PREMIUM_ROLES[element], premium=True)
+                    return await after.user.send('Thank you for becoming a premium supporter! Check out your shiney badges with `k!profile` and have fun with your new perks!')
 
-async def add_premium(member:discord.Member, s_id:int):
-  user = teams.find_one({'id': member.id})
-  badges = user['badges']
-  if not "premium" in user["badges"]:
-    badges.append('premium')
-  badges.append(premium[s_id])
-  teams.update_one({'id': member.id}, {'$set': {'badges': badges}})
-  try:
-    await member.send('Thank you for becoming a premium supporter! Check out your shiney badges with `k!profile` and have fun with your new perks!')
-  except:
-    pass
-  return
-  
+    @commands.Cog.listener()
+    async def on_command_error(self, ctx, error):
+        #This handels the k!bug cooldown
+        if isinstance(error, commands.CommandOnCooldown):
+            m, s = divmod(round(ctx.command.get_cooldown_retry_after(ctx)), 60)
+
+            return await ctx.send(f'Wait {m:02d} minutes and {s:02d} seconds before using the command again, thank you for helping to improve killua :3')
+
+        if isinstance(error, commands.BotMissingPermissions):
+            return await ctx.send(f"I don\'t have the required permissions to use this command! (`{', '.join(error.missing_perms)}`)")
+
+        if isinstance(error, commands.MissingPermissions):
+            return await ctx.send(f"You don\'t have the required permissions to use this command! (`', '.join({error.missing_perms})`)")
+        
+        if isinstance(error, commands.MissingRequiredArgument):
+            return await ctx.send(f"Seems like you missed a required argument for this command: `{error.param}`")
+
+        guild = ctx.guild.id if ctx.guild else "dm channel with "+ str(ctx.author.id)
+        command = ctx.command.name if ctx.command else "Error didn't occur during a command"
+        print('------------------------------------------')
+        print(f'An error occured\nGuild id: {guild}\nCommand name: {command}\nError: {error}')
+        print('------------------------------------------')
+
 Cog = Events
 
-
 def setup(client):
-  client.add_cog(Events(client))
+    client.add_cog(Events(client))
